提交修改更新病人列表刷新问题

+++ CareMobile/MendPatiendViewController.m
@@ -81,7 +81,7 @@
     
     NSString *str = [strDate substringWithRange:NSMakeRange(0, 7)];
     
-    NSString *strEnd = [NSString stringWithFormat:@"%@-%d",str,num];
+    NSString *strEnd = [NSString stringWithFormat:@"%@-%ld",str,(long)num];
     
     [_dateBtn setTitle:strEnd forState:0];
